+++ CHANGELOG.md
@@ -16,7 +16,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 ## Unreleased
 
 ### Added
-
+- DownStreamCacheControl middleware, which sets the `Edge-Control: no-store` header pages use csrf_token.
 
 ### Changed
 - Fixed issue surrounding table link download / external icons not appearing.